[CHORE/#77] enum class 분리
