+++ feature/src/main/java/com/terning/feature/home/home/model/SortBy.kt
@@ -1,0 +1,9 @@
+package com.terning.feature.home.home.model
+
+enum class SortBy(val sortBy: String) {
+    EARLIEST("deadlineSoon"),
+    SHORTEST("shortestDuration"),
+    LONGEST("longestDuration"),
+    SCRAP("mostScrapped"),
+    VIEW_COUNT("mostViewed"),
+}
